Move slider into common package
