+++ src/main/java/pl/ismop/web/client/widgets/common/slider/SliderView.ui.xml
