+++ src/main/java/pl/ismop/web/client/widgets/monitoring/fibre/FibrePresenter.java
@@ -20,7 +20,7 @@ import pl.ismop.web.client.dap.section.Section;
 import pl.ismop.web.client.error.ErrorDetails;
 import pl.ismop.web.client.widgets.common.map.MapPresenter;
 import pl.ismop.web.client.widgets.monitoring.fibre.IFibreView.IFibrePresenter;
-import pl.ismop.web.client.widgets.slider.SliderPresenter;
+import pl.ismop.web.client.widgets.common.slider.SliderPresenter;
 
 import java.util.*;
