+++ src/main/java/pl/ismop/web/client/widgets/root/RootPresenter.java
@@ -9,7 +9,6 @@ import pl.ismop.web.client.widgets.analysis.sidepanel.AnalysisSidePanelPresenter
 import pl.ismop.web.client.widgets.monitoring.mapnavigator.LeveeNavigatorPresenter;
 import pl.ismop.web.client.widgets.monitoring.sidepanel.MonitoringSidePanelPresenter;
 import pl.ismop.web.client.widgets.root.IRootPanelView.IRootPresenter;
-import pl.ismop.web.client.widgets.slider.SliderPresenter;
 
 @Presenter(view = RootPanel.class)
 public class RootPresenter extends BasePresenter<IRootPanelView, MainEventBus> implements IRootPresenter{
